+++ src/app/cart/model/cart.model.ts
@@ -1,0 +1,9 @@
+import { CartItem } from './cart-item.model';
+
+export class Cart {
+    constructor(
+        public items: CartItem[] = [],
+        public totalQuantity = 0,
+        public totalPrice = 0
+    ) {}
+}