+++ src/app/cart/services/cart.service.ts
@@ -2,45 +2,68 @@ import { Subject } from 'rxjs';
 
 import { Product } from '../../product/model/product.model';
 import { CartItem } from '../model/cart-item.model';
+import { Cart } from '../model/cart.model';
 
 export class CartService {
-  cartItems$ = new Subject();
+  cart$ = new Subject();
 
-  private cartContents: CartItem[] = [];
+  private cartItems: CartItem[] = [];
+  private totalQuantity = 0;
+  private totalPrice = 0;
 
   constructor() {}
 
   // Add to cart 1 item
-  addToCart(product: Product) {
+  addToCart(product: Product, quantity = 1) {
     const cartIndex = this.getItemIndex(product);
 
     if (cartIndex === -1) {
-      this.cartContents.push(new CartItem(product, 1));
+      this.cartItems.push(new CartItem(product, quantity));
     } else {
-      this.cartContents[cartIndex].quantity += 1;
+      this.cartItems[cartIndex].quantity += quantity;
     }
 
-    this.cartItems$.next(this.cartContents);
+    this.calculateTotals();
+    this.cart$.next(new Cart(this.cartItems, this.totalQuantity, this.totalPrice));
   }
 
   // Remove from cart 1 item
-  removeFromCart(product: Product) {
+  removeFromCart(product: Product, quantity?: number) {
     const cartIndex = this.getItemIndex(product);
 
     if (cartIndex === -1) {
       return;
     }
 
-    if (this.cartContents[cartIndex].quantity > 1) {
-      this.cartContents[cartIndex].quantity -= 1;
+    if (!quantity || quantity >= this.cartItems[cartIndex].quantity) {
+      // Remove items entirely
+      this.cartItems.splice(cartIndex, 1);
     } else {
-      this.cartContents.splice(cartIndex, 1);
+      this.cartItems[cartIndex].quantity -= quantity;
     }
 
-    this.cartItems$.next(this.cartContents);
+    this.calculateTotals();
+    this.cart$.next(new Cart(this.cartItems, this.totalQuantity, this.totalPrice));
+  }
+
+  clearCart() {
+    this.cartItems = [];
+
+    this.calculateTotals();
+    this.cart$.next(new Cart(this.cartItems, this.totalQuantity, this.totalPrice));
   }
 
   private getItemIndex(product: Product): number {
-    return this.cartContents.findIndex((cartItem) => cartItem.product.name === product.name);
+    return this.cartItems.findIndex((cartItem) => cartItem.product.name === product.name);
+  }
+
+  private calculateTotals() {
+    this.totalQuantity = this.cartItems.reduce((value, item) => {
+      return value + item.quantity;
+    }, 0);
+
+    this.totalPrice = this.cartItems.reduce((value, item) => {
+      return value + (item.quantity * item.product.price);
+    }, 0);
   }
 }