+++ changelog.md
@@ -19,3 +19,11 @@
     - Add subscriptions and handlers for Cart and Product services to CartList and ProductList components
     - Display available quantity of items and total quantity and price of items in cart
 - Add Highlight directive
+
+
+## Task 3
+
+- Update CartService to keep track of total number of items in cart and total price
+- Create ConfigOptions, Constants, Generator and LocalStorage services in Core module
+- Create Demo component to use this services
+- Create Clicker directive